use floats (our line height is now 22.5px)

+++ js/flowingColumns.mjs
@@ -15,12 +15,12 @@ class FlowingColumns {
     #spacer = this.#column?.querySelector('#spacer');
     #lastParagraph = this.#column?.lastElementChild;
     #paragraphLineHeight = this.#lastParagraph
-        ? parseInt(window.getComputedStyle(this.#lastParagraph).getPropertyValue('line-height'))
+        ? parseFloat(window.getComputedStyle(this.#lastParagraph).getPropertyValue('line-height'))
         : null;
     #firstHeader = this.#column?.querySelector('h3');
     #headerHeight = this.#firstHeader?.getBoundingClientRect().height;
     #baseOffset = this.#firstHeader
-        ? parseInt(window.getComputedStyle(this.#firstHeader).getPropertyValue('padding-top'))
+        ? parseFloat(window.getComputedStyle(this.#firstHeader).getPropertyValue('padding-top'))
         : null;
 
     #offset = 0;
@@ -30,7 +30,7 @@ class FlowingColumns {
     #matchesMedia = this.#mediaQuery.matches;
 
     #diffMoreThan = (x, y, threshold) => Math.abs(x - y) > threshold;
-    #roundNearest = (value, by) => by * Math.round(value / by);
+    #roundNearest = (value, interval) => interval * Math.round(value / interval);
 
     #flowColumns = () => {
         // offset of page readings
@@ -59,7 +59,7 @@ class FlowingColumns {
 
         }
 
-        // add an extra margin on top: with a maximum of the height of the biggest child and a minimum of half a line
+        // add an extra margin on top
         const
             safetyMargin = this.#headerHeight / 2,
             heightNew = Math.max(1, calculatedHeight + safetyMargin);
